refactor: extract message sending logic
