+++ src/types.ts
@@ -22,7 +22,6 @@ export interface Book {
 }
 
 export interface BackgroundResponse {
-	success?: boolean;
 	data?: PageObjectResponse | GetDatabaseResponse;
 	error?: ExtensionError;
 }
@@ -56,7 +55,7 @@ export type ExtensionError =
 			message: "Sorry, D2N was unable to retrieve information about the book from this page. Please make sure you are on a valid book page and try again";
 	  }
 	| {
-		name: "Token&IDNotExistError";
-			message: "Sorry, You have to provide Notion Secret Token and Database ID to use D2N"
+			name: "Token&IDNotExistError";
+			message: "Sorry, You have to provide Notion Secret Token and Database ID to use D2N";
 	  }
 	| NotionClientError;
